requirements.txt
tensorflow==2.14.0
keras==2.14.0
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
seaborn==0.12.2
pillow==10.0.0
flask==2.3.0
flask-cors==4.0.0